requirements.txt
flask
flask-sqlalchemy
flask-session
python-dotenv
bcrypt
flask_cors
flask_wtf